pyproject.toml
[tool.poetry]
name = "currency-conversion"
version = "0.1.0"
description = ""
authors = ["Anatolii <a.akinshin@satvaspace.com>"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["all"], version = "^0.109.2"}
fastapi-sqlalchemy = "^0.2.1"
alembic = "^1.13.1"
gunicorn = "^21.2.0"
psycopg = {extras = ["binary", "pool"], version = "^3.1.18"}
psycopg2-binary = "^2.9.9"
aiohttp = "^3.9.3"

[tool.poetry.dev-dependencies]

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.1"
ruff = "^0.2.2"
coverage = "^7.4.2"
pytest-cov = "^4.1.0"
pytest-asyncio = "^0.23.5"
factory-boy = "^3.3.0"
httpx = "^0.27.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
